tratamento erro controller

+++ src/main/java/br/com/itau/geradornotafiscal/service/CalculadoraFreteService.java
@@ -2,6 +2,8 @@ package br.com.itau.geradornotafiscal.service;
 
 import br.com.itau.geradornotafiscal.model.Pedido;
 
+import java.math.BigDecimal;
+
 public interface CalculadoraFreteService {
-    double calcularFrete(Pedido pedido);
+    BigDecimal calcularFrete(Pedido pedido);
 }
