+++ src/test/java/br/com/itau/geradornotafiscal/service/impl/GeradorNotaFiscalServiceImplTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.*;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 
@@ -53,8 +54,8 @@ public class GeradorNotaFiscalServiceImplTest {
     @BeforeEach
     void setUp() {
         pedido = new Pedido();
-        pedido.setValorTotalItens(100.0);
-        pedido.setValorFrete(10.0);
+        pedido.setValorTotalItens(BigDecimal.valueOf(100.0));
+        pedido.setValorFrete(BigDecimal.valueOf(10.0));
         Destinatario destinatario = new Destinatario();
         destinatario.setTipoPessoa(TipoPessoa.FISICA);
         Endereco endereco = new Endereco();
@@ -69,10 +70,10 @@ public class GeradorNotaFiscalServiceImplTest {
 
     @Test
     void deveGerarNotaFiscalCorretamente() {
-        when(calculadoraAliquotaService.calcularAliquota(pedido)).thenReturn(0.2);
-        when(calculadoraFreteService.calcularFrete(pedido)).thenReturn(50.0);
-        when(itemNotaFiscalFactory.gerarItensNotaFiscal(pedido, 0.2)).thenReturn(Collections.singletonList(item));
-        when(notaFiscalFactory.criarNotaFiscal(pedido, Collections.singletonList(item), 50.0)).thenReturn(notaFiscalEsperada);
+        when(calculadoraAliquotaService.calcularAliquota(pedido)).thenReturn(BigDecimal.valueOf(0.2));
+        when(calculadoraFreteService.calcularFrete(pedido)).thenReturn(BigDecimal.valueOf(50.0));
+        when(itemNotaFiscalFactory.gerarItensNotaFiscal(pedido, BigDecimal.valueOf(0.2))).thenReturn(Collections.singletonList(item));
+        when(notaFiscalFactory.criarNotaFiscal(pedido, Collections.singletonList(item), BigDecimal.valueOf(50.0))).thenReturn(notaFiscalEsperada);
 
         NotaFiscal resultado = geradorNotaFiscalService.gerarNotaFiscal(pedido);
 
@@ -82,12 +83,12 @@ public class GeradorNotaFiscalServiceImplTest {
 
     @Test
     void deveGerarNotaFiscalComFreteZero() {
-        when(calculadoraAliquotaService.calcularAliquota(pedido)).thenReturn(0.1);
-        when(calculadoraFreteService.calcularFrete(pedido)).thenReturn(0.0);
+        when(calculadoraAliquotaService.calcularAliquota(pedido)).thenReturn(BigDecimal.valueOf(0.1));
+        when(calculadoraFreteService.calcularFrete(pedido)).thenReturn(BigDecimal.valueOf(0.0));
         NotaFiscal notaFiscalMock = mock(NotaFiscal.class);
         List<ItemNotaFiscal> itens = Collections.singletonList(mock(ItemNotaFiscal.class));
-        when(itemNotaFiscalFactory.gerarItensNotaFiscal(pedido, 0.1)).thenReturn(itens);
-        when(notaFiscalFactory.criarNotaFiscal(pedido, itens, 0.0)).thenReturn(notaFiscalMock);
+        when(itemNotaFiscalFactory.gerarItensNotaFiscal(pedido, BigDecimal.valueOf(0.1))).thenReturn(itens);
+        when(notaFiscalFactory.criarNotaFiscal(pedido, itens, BigDecimal.valueOf(0.0))).thenReturn(notaFiscalMock);
 
         NotaFiscal resultado = geradorNotaFiscalService.gerarNotaFiscal(pedido);
 
@@ -147,7 +148,7 @@ public class GeradorNotaFiscalServiceImplTest {
 
     @Test
     void deveLancarExcecaoQuandoValorTotalItensForZero() {
-        pedido.setValorTotalItens(0.0);
+        pedido.setValorTotalItens(BigDecimal.valueOf(0.0));
 
         PedidoInvalidoException exception = assertThrows(PedidoInvalidoException.class, () ->
                 geradorNotaFiscalService.gerarNotaFiscal(pedido)
@@ -157,7 +158,7 @@ public class GeradorNotaFiscalServiceImplTest {
 
     @Test
     void deveLancarExcecaoQuandoValorFreteForNegativo() {
-        pedido.setValorFrete(-10.0);
+        pedido.setValorFrete(BigDecimal.valueOf(-10.0));
 
         PedidoInvalidoException exception = assertThrows(PedidoInvalidoException.class, () ->
                 geradorNotaFiscalService.gerarNotaFiscal(pedido)
@@ -204,10 +205,10 @@ public class GeradorNotaFiscalServiceImplTest {
         NotaFiscal notaFiscalMock = mock(NotaFiscal.class);
         List<ItemNotaFiscal> itens = Collections.singletonList(mock(ItemNotaFiscal.class));
 
-        when(calculadoraAliquotaService.calcularAliquota(pedido)).thenReturn(0.1);
-        when(calculadoraFreteService.calcularFrete(pedido)).thenReturn(15.0);
-        when(itemNotaFiscalFactory.gerarItensNotaFiscal(pedido, 0.1)).thenReturn(itens);
-        when(notaFiscalFactory.criarNotaFiscal(pedido, itens, 15.0)).thenReturn(notaFiscalMock);
+        when(calculadoraAliquotaService.calcularAliquota(pedido)).thenReturn(BigDecimal.valueOf(0.1));
+        when(calculadoraFreteService.calcularFrete(pedido)).thenReturn(BigDecimal.valueOf(15.0));
+        when(itemNotaFiscalFactory.gerarItensNotaFiscal(pedido, BigDecimal.valueOf(0.1))).thenReturn(itens);
+        when(notaFiscalFactory.criarNotaFiscal(pedido, itens, BigDecimal.valueOf(15.0))).thenReturn(notaFiscalMock);
 
         if (servicoMock == estoqueService) {
             doThrow(new RuntimeException(erro)).when(estoqueService).enviarNotaFiscalParaBaixaEstoque(notaFiscalMock);
